[Qt] Don't override default QApplication font for DRT and WTR
Revision 95526 moved the relevant code below the QApplication
constructor, to fix a potential crash, and that made the code
take effect, breaking 6 tests.

We should rely on the QWebSettings to set default fonts, and
this QApplication::setFont() is a relic from when we didn't
have that option, so we remove the code.

Reviewed by Csaba Osztrogonác.

* DumpRenderTree/qt/main.cpp:
* WebKitTestRunner/InjectedBundle/qt/ActivateFontsQt.cpp:

Identifier: 84388@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@95541 268f45cc-cd09-0410-ab3c-d52691b4dbfc