[GTK] Fix webkit2 unit tests in debug builds https://bugs.webkit.org/show_bug.cgi?id=69006
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2011-09-28
Reviewed by Martin Robinson.

We are currently using the WorQueue name as thread name which in
this moment can be com.apple.WebKit.ProcessLauncher or
com.apple.CoreIPC.ReceiveQueue. Both names are longer than 31
characters which is the limit of Visual Studio for thread
names. When log is enabled createThread() will assert instead of
truncate the name, so we need to make sure we don't use a name
longer than 31 characters.

* Platform/gtk/WorkQueueGtk.cpp:
(WorkQueue::platformInitialize):

Identifier: 85006@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@96237 268f45cc-cd09-0410-ab3c-d52691b4dbfc