[Freetype] Some text in Planet GNOME renders in the wrong place https://bugs.webkit.org/show_bug.cgi?id=69099
Reviewed by Dirk Schulze.

Source/WebCore:

Test: platform/gtk/fonts/synthetic-oblique-positioning.html

Fold the oblique transform into the font matrix itself rather than transforming
the CTM of the context at render time. Not only does this fix the issue, it
prevents unnecessary work on every paint.

* platform/graphics/cairo/FontCairo.cpp:
No longer set the synthetic oblique transformation matrix on the context
when rendering the text. Instead it is now folded into the TM of the font
itself.
* platform/graphics/freetype/FontPlatformDataFreeType.cpp:
(WebCore::FontPlatformData::initializeWithFontFace): Fold the oblique transform into the scaled font.
* platform/graphics/win/FontPlatformDataCairoWin.cpp:
(WebCore::FontPlatformData::FontPlatformData): Ditto.

LayoutTests:

Add a test which exercises this issue. It seems the problem grows worse
as the y component of the text position increases, so position the text
about halfway down the page.

* platform/gtk/fonts/synthetic-oblique-positioning-expected.png: Added.
* platform/gtk/fonts/synthetic-oblique-positioning-expected.txt: Added.
* platform/gtk/fonts/synthetic-oblique-positioning.html: Added.

Identifier: 85141@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@96378 268f45cc-cd09-0410-ab3c-d52691b4dbfc